backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3

# Database
pymongo==4.6.0
dnspython==2.4.2
pydantic==2.5.0

# AI & Machine Learning
tensorflow==2.15.0
opencv-python==4.8.1.78
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
Pillow==10.1.0

# Image Processing
Pillow==10.1.0
imageio==2.33.0

# Natural Language Processing
nltk==3.8.1
textblob==0.17.1

# API & Serialization
marshmallow==3.20.1
python-dotenv==1.0.0
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.1
